Refactor: Introduce BaseWriter to reduce code duplication
- Created a BaseWriter class to consolidate common logic from ProtoDataStoreWriter and PreferencesWriter. Moved the writePackage function to BaseWriter to eliminate code duplication and improve maintainability.
